+++ apps/native-component-list/src/screens/UI/UIScreen.tsx
@@ -1,13 +1,20 @@
-import { View, Text } from 'react-native';
+import { optionalRequire } from '../../navigation/routeBuilder';
+import ComponentListScreen, { componentScreensToListElements } from '../ComponentListScreen';
 
-export const UIScreens = [];
+export const UIScreens = [
+  {
+    name: 'Community SegmentedControl replacement',
+    route: 'ui/community-segmented-control',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./CommunitySegmentedControlScreen'));
+    },
+  },
+];
 
 export default function UIScreen() {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>Expo UI is only available on Android and iOS currently.</Text>
-    </View>
-  );
+  const apis = componentScreensToListElements(UIScreens);
+  return <ComponentListScreen apis={apis} sort />;
 }
 
 UIScreen.navigationOptions = {
