--- packages/expo-ui/src/community/segmented-control/SegmentedControl.web.tsx
@@ -1,5 +1,0 @@
-import type { SegmentedControlProps } from './types';
-
-export function SegmentedControl(_props: SegmentedControlProps) {
-  return null;
-}
