+++ packages/expo-ui/src/community/segmented-control/types.tsx
@@ -43,7 +43,7 @@ export type SegmentedControlProps = {
   onValueChange?: (value: string) => void;
   /**
    * Accent color of the control.
-   * @platform android
+   * @platform android, web
    */
   tintColor?: string;
   /**
